backend/requirements.txt
# Core
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.25
alembic>=1.13.1
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
redis>=5.0.1
celery>=5.3.6
pydantic>=2.10.4
pydantic-settings>=2.1.0
python-multipart>=0.0.6
websockets>=12.0
httpx>=0.27.2

# AI/Browser
browser-use>=0.1.40
langchain>=1.2.0
langchain-core>=1.2.0
langchain-google-genai>=4.0.0
langchain-openai>=1.0.0
langchain-anthropic>=1.0.0
playwright>=1.41.0

# Auth
bcrypt>=4.0.0
python-jose[cryptography]>=3.3.0
email-validator>=2.0.0

# API Testing
jsonpath-ng>=1.6.0
jsonschema>=4.21.0
pyyaml>=6.0.1
